Fix issue on form/modal not closing on submit:
1. getProfile Function in Account:
Added a showLoading parameter to control whether the spinner is displayed.
This prevents the spinner from interfering with the modal when refreshing the profile data after a successful update.

2.Modal in ProfileCard:

The ProfileUpdateForm now calls getProfile(false) to refresh the profile data without showing the spinner, ensuring the modal remains open and the success message is displayed.

3. Also added SEO for Account